Attribute Macro gd_rehearse::bench::gdbench

#[gdbench]

Benchmark for gdext classes and functions integrated with Godot.

This macro transforms and registers the annotated function for further usage by GdTestRunner within a Godot test scene. When the runner enters the scene, it executes all qualified benchmarks.

A function annotated with #[gdbench] must:

Have a return value.

Have no parameters or only a singular BenchContext.

Every benchmark is executed 200 times for a warm-up, followed by 501 additional runs to assess runtime (an odd number of runs for easy median extraction). Minimum and median run times will be displayed.

Attributes

An attribute-less macro will make the benchmark run, but several attributes are available for better customizability, especially when working on specific attributes and creating more narrowly-focused test runner scenes.

skip: Skips the benchmark during execution.

focus: Forces a focused run, in which only benchmarks annotated with focus will be executed.

keyword: A specific keyword that will be picked up by the runner. The benchmark runs only if the runner has the same keyword specified.

scene_path: Godot path to the scene. If specified, given benchmark will only run if runner’s scene path is the same.

repeat: Specifies the number of internal repeats the benchmark should undergo. By default, the function executes 100 times within every run.

setup: Optional function that will be executed before benchmark execution, to set up the scene for benchmarks.

cleanup: Optional function that will be executed after benchmark execution, to clean up after benchmarks. Rarely needed, as when setup is present, the default cleanup function should always clean up efficiently.

Examples

use gd_rehearse::bench::*;
use godot::prelude::*;

// Causes a focus run during which only the focused benchmarks will be executed, but only with
// `my bench` as a keyword in the runner.
#[gdbench(focus, keyword = "my bench")]
fn focused_bench() -> i32 {
    1337
}

// Will be skipped during the default run.
#[gdbench(skip)]
fn skipped_test() -> i32 {
    231 + 312
}

// Can access the `GdTestRunner` scene_tree.
#[gdbench]
fn bench_with_ctx(ctx: &BenchContext) -> InstanceId {
    ctx.scene_tree().instance_id()
}

// Will only run in `res://special_cases.tscn` scene.
#[gdbench(scene_path="res://special_cases.tscn")]
fn bench_with_path(ctx: &BenchContext) -> Variant {
    let test_node = ctx.scene_tree().get_node_or_null("SomeTestNode".into()).expect("Can't get node");
    let variant = test_node.get("property_should_be_here".into());
    assert!(!variant.is_nil());
    variant
}