gd_rehearse::bench

Struct BenchContext

pub struct BenchContext { /* private fields */ }

Context for Rust Benchmarking case.

Object allowing access to the scene tree during #[gdbench] benchmarking function execution. Provides simple, general access to the scene tree relative to GdTestRunner in which the banchmark takes place, as well as some utility functions for retrieving nodes present in the scene or set up before benchmark with setup function.

Usage of specialized node-retrieving functions is recommended - some overhead will be still recorded, but the duration which is crucial in benchmarking will be adjusted mostly.

Examples

use godot::prelude::*;
use gd_rehearse::bench::*;

fn setup_function(ctx: &mut BenchContext) {
   let node = Node::new_alloc();
   ctx.setup_add_node(node, "OnSetup");
}

#[gdbench(setup=setup_function, scene_path="res://scene_with_specific.tscn")]
fn with_setup(ctx: &BenchContext) -> bool {
   let _from_setup = ctx.get_setup_node("OnSetup");
   let _from_scene = ctx.get_node("SceneSpecific");
   true
}

Implementations

impl BenchContext

pub fn remove_all_added_nodes(&mut self)

Removes all nodes added during the setup procedure.

This method is called during default cleanup procedure, and needs to be called also when implementing some custom cleanup is required.

pub fn remove_added_node(&mut self, name: impl Into<GString>)

Removes single node added during the setup procedure.

For usage in custom cleanup procedure, if you need to remove the nodes in specific order. All nodes need to be cleaned up during cleanup.

pub fn setup_add_node(&mut self, node: Gd<Node>, name: impl Into<GString>)

Add node to scene in which the benchmark will be processed.

This method should be called only during setup procedure for a benchmark, to prepare some objects which shouldnโt be generated within the benchmark run itself.

pub fn get_setup_node(&self, name: impl Into<GString>) -> Gd<Node>

Gets node from current benchmark context that was set up.

Panics

If no node with name was set up during setup function.

pub fn get_setup_node_as<T>(&self, name: impl Into<GString>) -> Gd<T>
where T: Inherits<Node>,

Gets node from current benchmark context that was set up, upcasted to T.

Panics

If no node with name was set up during setup function, or cannot be upcasted to T.