Struct gd_rehearse::GdTestRunner

pub struct GdTestRunner { /* private fields */ }

Tests and benchmark runner for custom Godot classes created using the gdext crate.

Runs functions annotated with #[gditest] and #[gdbench] macros, facilitating the testing and benchmarking of methods and functions that involve calls between Rust and Godot. To utilize it, create a scene in the Godot project associated with your gdext-based GDExtension and run the scene either from the command line or directly from the Godot editor.

Godot Properties

GdTestRunner exposes some settable Godot-exported properties that customize its behavior:

run_tests: If set, functions annotated with #[gditest] will be executed. Defaults to true.

run_benchmarks: If set, functions annotated with #[gdbench] will be executed. Defaults to true. If run_tests is also true, benchmarks will only be executed if all tests pass successfully.

test_keyword: If set, only tests and benchmarks with the same keyword specified will be executed. Defaults to an empty string, meaning that only tests and benchmarks without a keyword set will be executed. It takes precedence over focus and filters—they will be assessed, but only in the context of this keyword.

ignore_keywords: If set, all tests and benchmarks will be executed regardless of their set keyword.

disallow_focus: If set, the focus attribute of tests and benchmarks will be ignored.

disallow_skip: If set, the skip attribute of tests and benchmarks will be ignored.

test_filters: An array of strings tested against the names of tests and benchmarks. Those with names containing at least one of the specified filters will be executed.

Command Line Arguments

GdTestRunner is also suitable for running from the command line, as part of the Godot binary execution in headless mode:

godot_executable --path path/to/godot/project [scene path] --headless -- [optional arguments]

scene_path: An optional path to the scene with the GdTestRunner object. Not needed if the testing scene is the main scene of the project.

optional arguments: Options such as:

--rust-test or --rust-benchmarks: If at least one is selected, overwrites the analogous properties and runs only the specified element.

--disallow-focus or --allow-focus: Overwrites the disallow_focus property.

--disallow-skip or --disallow-focus: Overwrites the disallow_skip property.

--mute-keyword or --keyword=my_keyword: Either mutes the test_keyword property or replaces it with the specified one.

--ignore-keywords: Replaces the ignore_keywords property.

--mute-filters or --filters=[filter1,filter2]: Either mutes the test_filters property or replaces it with the specified filters.

Implementations

impl GdTestRunner

pub fn get_run_tests(&self) -> <bool as Var>::Intermediate

pub fn set_run_tests(&mut self, run_tests: <bool as Var>::Intermediate)

pub fn get_run_benchmarks(&self) -> <bool as Var>::Intermediate

pub fn set_run_benchmarks( &mut self, run_benchmarks: <bool as Var>::Intermediate )

pub fn get_disallow_focus(&self) -> <bool as Var>::Intermediate

pub fn set_disallow_focus( &mut self, disallow_focus: <bool as Var>::Intermediate )

pub fn get_disallow_skip(&self) -> <bool as Var>::Intermediate

pub fn set_disallow_skip(&mut self, disallow_skip: <bool as Var>::Intermediate)

pub fn get_test_keyword(&self) -> <GString as Var>::Intermediate

pub fn set_test_keyword(&mut self, test_keyword: <GString as Var>::Intermediate)

pub fn get_ignore_keywords(&self) -> <bool as Var>::Intermediate

pub fn set_ignore_keywords( &mut self, ignore_keywords: <bool as Var>::Intermediate )

pub fn get_test_filters(&self) -> <PackedStringArray as Var>::Intermediate

pub fn set_test_filters( &mut self, test_filters: <PackedStringArray as Var>::Intermediate )