Crate gd_rehearse

Framework for testing and benchmarking godot-rust-dependent code.

When using the standard #[test] Rust macro in a Godot project, tests may fail if they involve objects requiring the Godot executable to run. godot-test addresses this limitation by introducing the #[gditest] and #[gdbench] macros. These can be used to annotate functions intended for testing or benchmarking your code, and the crate provides the GdTestRunner GodotClass for executing them within a Godot scene.

Modules

bench

Contains all symbols necessary to use #[gdbench] macro.

itest

Contains all symbols necessary to use #[gditest] macro.

Structs

GdTestRunner

Tests and benchmark runner for custom Godot classes created using the gdext crate.