gd_rehearse::bench

Trait CaseContext

pub trait CaseContext {
    // Required method
    fn scene_tree(&self) -> &Gd<Node>;

    // Provided methods
    fn get_node(&self, path: impl AsArg<NodePath>) -> Gd<Node> { ... }
    fn get_node_as<T>(&self, path: impl AsArg<NodePath>) -> Gd<T>
       where T: Inherits<Node> { ... }
}

Required Methods

fn scene_tree(&self) -> &Gd<Node>

Get access to current scene tree.

Access is provided relatively to the GdTestRunner position in the current scene (it’s recommended that it is a root node of the scene).

Use it if you need to do something else besides getting access to some specific node, especially during benchmarking. For node retrieval you can use CaseContext::get_node and CaseContext::get_node_as.

Provided Methods

fn get_node(&self, path: impl AsArg<NodePath>) -> Gd<Node>

Gets node located at path, relatively to GdTestRunner.

Panics

Panics if no node is present at the path.

fn get_node_as<T>(&self, path: impl AsArg<NodePath>) -> Gd<T>
where T: Inherits<Node>,

Gets node located at path, relatively to GdTestRunner, casted to T.

Panics

Panics if no node is present at the path, or cannot be casted to T.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.