gd_rehearse::itest

Attribute Macro gditest

#[gditest]

Integration test between Godot and Rust.

Similar to #[test], but converts the function into an integration test between Godot and Rust.

It transforms and registers the annotated function for further usage by GdTestRunner in some Godot test scene. When the runner enters the scene, it will run all qualified tests.

A function annotated with #[gditest] needs to:

Have no return values.

Have no parameters or only a singular TestContext.

Attributes

An attribute-less macro will make the tests run, but some attributes are available for better customizability, especially when working on specific attributes and creating more narrow test runner scenes.

skip: Skips the test during run.

focus: Forces focus run, in which only tests annotated with focus will be run.

keyword: A specific keyword that will be picked up by the runner, and the test will be run only if the runner has the same keyword specified.

scene_path: Godot path to the scene. If specified, given benchmark will only run if runner’s scene path is the same.

Examples

use gd_rehearse::itest::*;

// Causes a focus run during which only the focused tests will be executed, but only with
// `my test` as a keyword in the runner.
#[gditest(focus, keyword = "my test")]
fn focused_test() {
    assert!(true);
}

// Will be skipped during the default run.
#[gditest(skip)]
fn skipped_test() {
    let test = 1 + 1;
    assert_eq!(test, 1);
}

// Can access the `GdTestRunner` scene_tree.
#[gditest]
fn test_with_ctx(ctx: &TestContext) {
    ctx.scene_tree().instance_id();
}

// Will only run in `res://special_cases.tscn` scene.
#[gditest(scene_path="res://special_cases.tscn")]
fn test_with_path(ctx: &TestContext) {
    let test_node = ctx.scene_tree().get_node_or_null("SomeTestNode");
    assert!(test_node.is_some());
    assert!(!test_node.unwrap().get("property_should_be_here").is_nil());
 
    // Shorthand for getting node by name.
    let test_node = ctx.get_node("SomeTestNode");
    assert!(!test_node.get("property_should_be_here").is_nil());
}